fix date and time
